Update 201.01-lesson-originsofmoney.md
I'm not really sure talking about the coincidence of double want to be the way to explain barter. 

maybe this could be mentioned, but this is very abstract concept. 

Have you ever looked at this [book for kids](https://www.amazon.com/Bitcoin-Money-Tale-Bitville-Discovering/dp/0578490676)?

it explains barter quite well with limonade, law mowing and drawing. it's much more practical/tangible than the way it is been described here.